analysis/requirements.txt
yfinance>=0.2.31
pandas>=2.1.0
pandas-ta>=0.3.14b1
pymongo>=4.6.0
redis>=5.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
numpy>=1.24.0
flask>=3.0.0
